+++ rector.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 use Rector\CodeQuality\Rector\Class_\InlineConstructorDefaultToPropertyRector;
 use Rector\Config\RectorConfig;
-use Rector\Php71\Rector\FuncCall\CountOnNullRector;
 use Rector\Php73\Rector\String_\SensitiveHereNowDocRector;
 use Rector\Php74\Rector\Closure\ClosureToArrowFunctionRector;
 use Rector\Set\ValueObject\LevelSetList;
@@ -12,7 +11,10 @@ use Rector\Set\ValueObject\LevelSetList;
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->paths([
         __DIR__ . '/src',
-        __DIR__ . '/tests',
+        /**
+         * Disabled ./tests directory due to different branches with main package when testing
+         */
+        // __DIR__ . '/tests',
     ]);
 
     // register a single rule
@@ -26,8 +28,5 @@ return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->skip([
         ClosureToArrowFunctionRector::class,
         SensitiveHereNowDocRector::class,
-        CountOnNullRector::class => [
-            'src/Command.php',
-        ],
     ]);
 };