+++ tests/CommandTest.php
@@ -65,7 +65,7 @@ final class CommandTest extends CommonCommandTest
     public function testBatchInsert(
         string $table,
         array $columns,
-        array $values,
+        iterable $values,
         string $expected,
         array $expectedParams = [],
         int $insertedRow = 1
@@ -527,9 +527,10 @@ final class CommandTest extends CommonCommandTest
         array $columns,
         array|string $conditions,
         array $params,
-        string $expected
+        array $expectedValues,
+        int $expectedCount,
     ): void {
-        parent::testUpdate($table, $columns, $conditions, $params, $expected);
+        parent::testUpdate($table, $columns, $conditions, $params, $expectedValues, $expectedCount);
     }
 
     /**